:recycle: Refactor: Oauth 테이블 분리
<footer>
- 관련: #180
