+++ storage/db-mysql-v2/src/main/java/com/namo/spring/db/mysql/common/converter/ProviderConverter.java
@@ -1,6 +1,6 @@
 package com.namo.spring.db.mysql.common.converter;
 
-import com.namo.spring.db.mysql.domains.user.type.Provider;
+import com.namo.spring.db.mysql.domains.oauth.type.Provider;
 
 public class ProviderConverter extends AbstractEnumAttributeConverter<Provider>{
 	private static final String ENUM_NAME = "소셜 로그인 제공자";
